+++ app/src/main/java/com/zype/android/ui/main/fragments/videos/VideosActivity.java
@@ -22,6 +22,7 @@ import com.android.billingclient.api.BillingClient;
 import com.android.billingclient.api.Purchase;
 import com.squareup.otto.Subscribe;
 import com.zype.android.Auth.AuthHelper;
+import com.zype.android.subscription.SubscriptionHelper;
 import com.zype.android.Billing.BillingManager;
 import com.zype.android.Billing.SubscriptionsHelper;
 import com.zype.android.DataRepository;
@@ -346,7 +347,7 @@ public class VideosActivity extends MainActivity implements ListView.OnItemClick
 
         selectedVideoId = holder.videoId;
         NavigationHelper navigationHelper = NavigationHelper.getInstance(this);
-        if (AuthHelper.isVideoUnlocked(this, holder.videoId, playlistId)) {
+        if (SubscriptionHelper.isVideoUnlocked(this, holder.videoId, playlistId)) {
             navigationHelper.switchToVideoDetailsScreen(this, holder.videoId, playlistId, false);
         }
         else {
